+++ client/src/pages/index.js
@@ -6,7 +6,7 @@ import ForgotPassword from './ForgotPassword';
 import ResetPassword from './ResetPassword';
 import Admin from './Admin';
 import MentalHeath from './OrgPublic/MentalHeath';
-import BudgetingPage from './BudgetingPage';
+import Budgeting from './OrgPublic/Budgeting';
 import SubSections from './OrgPublic/SubSections';
 
 export {
@@ -19,5 +19,5 @@ export {
   ResetPassword,
   MentalHeath,
   SubSections,
-  BudgetingPage,
+  Budgeting,
 };
